feat/#27: 모든 마커를 띄우는 showMarkers 정의 - 위젯 트리에서 임시 버튼 depth 수정
